Refactor context creation into separate definition files
