+++ src/contexts/AppContextDef.tsx
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { AppContextType } from './AppContext';
+
+export const AppContext = createContext<AppContextType | undefined>(undefined);
