+++ src/contexts/AuthContextDef.tsx
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { AuthContextType } from '../types/auth';
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
